+++ src/main/scala/com/colofabrix/scala/tado4s/api/HomeUserResponse.scala
@@ -58,7 +58,7 @@ final case class MobileDevices(
   name: String,
   id: Int,
   settings: Settings,
-  location: Location,
+  location: Option[Location],
   deviceMetadata: DeviceMetadata,
 ) derives Decoder